+++ pkg/ontap/client/zapi.go
@@ -29,7 +29,7 @@ func NewOntapZAPI(nodeConfig *config.NodeConfig) (c *OntapZAPI, err error) {
 			BasicAuthPassword: nodeConfig.Storage.CdotCredentials.Password,
 			SSLVerify:         false,
 			Debug:             false,
-			Timeout:           120 * time.Second,
+			Timeout:           300 * time.Second,
 			Version:           nodeConfig.Storage.CdotCredentials.ZapiVersion,
 		},
 	)